Move OSFindResource() and related methods to IPlugPaths.
This also removes ToNSString and CocoaAutoReleasePool
