+++ IGraphics/Platforms/IGraphicsWeb.cpp
@@ -413,38 +413,6 @@ ECursor IGraphicsWeb::SetMouseCursor(ECursor cursorType)
   return IGraphics::SetMouseCursor(cursorType);
 }
 
-EResourceLocation IGraphicsWeb::OSFindResource(const char* name, const char* type, WDL_String& result)
-{
-  if (CStringHasContents(name))
-  {
-    WDL_String plusSlash;
-    
-    bool foundResource = false;
-    
-    //TODO: OSFindResource is not sufficient here
-    
-    if(strcmp(type, "png") == 0) { //TODO: lowercase/uppercase png
-      plusSlash.SetFormatted(strlen("/resources/img/") + strlen(name) + 1, "/resources/img/%s", name);
-      foundResource = GetPreloadedImages().call<bool>("hasOwnProperty", std::string(plusSlash.Get()));
-    }
-    else if(strcmp(type, "ttf") == 0) { //TODO: lowercase/uppercase ttf
-      plusSlash.SetFormatted(strlen("/resources/fonts/") + strlen(name) + 1, "/resources/fonts/%s", name);
-      foundResource = true; // TODO: check ttf
-    }
-    else if(strcmp(type, "svg") == 0) { //TODO: lowercase/uppercase svg
-      plusSlash.SetFormatted(strlen("/resources/img/") + strlen(name) + 1, "/resources/img/%s", name);
-      foundResource = true; // TODO: check svg
-    }
-    
-    if(foundResource)
-    {
-      result.Set(plusSlash.Get());
-      return EResourceLocation::kAbsolutePath;
-    }
-  }
-  return EResourceLocation::kNotFound;
-}
-
 //static
 void IGraphicsWeb::OnMainLoopTimer()
 {
