+++ IGraphics/Platforms/IGraphicsWin.h
@@ -71,10 +71,6 @@ public:
 
   bool GetTextFromClipboard(WDL_String& str) override;
 
-  EResourceLocation OSFindResource(const char* name, const char* type, WDL_String& result) override;
-
-  const void* LoadWinResource(const char* resid, const char* resType, int& sizeInBytes) override;
-
 protected:
   IPopupMenu* CreatePlatformPopupMenu(IPopupMenu& menu, const IRECT& bounds, IControl* pCaller) override;
   void CreatePlatformTextEntry(IControl& control, const IText& text, const IRECT& bounds, const char* str) override;
