+++ IPlug/IPlugPaths.cpp
@@ -139,6 +139,62 @@ void INIPath(WDL_String& path, const char * pluginName)
   path.AppendFormatted(MAX_WIN32_PATH_LEN, "\\%s", pluginName);
 }
 
+EResourceLocation OSFindResource(const char* name, const char* type, WDL_String& result, const char*)
+{
+  if (CStringHasContents(name))
+  {
+    WDL_String search(name);
+    WDL_String typeUpper(type);
+
+    EnumResourceNames(mHInstance, _strupr(typeUpper.Get()), (ENUMRESNAMEPROC)EnumResNameProc, (LONG_PTR)&search);
+
+    if (strstr(search.Get(), "found: ") != 0)
+    {
+      result.SetFormatted(MAX_PATH, "\"%s\"", search.Get() + 7, search.GetLength() - 7); // 7 = strlen("found: ")
+      return EResourceLocation::kWinBinary;
+    }
+    else
+    {
+      if (PathFileExists(name))
+      {
+        result.Set(name);
+        return EResourceLocation::kAbsolutePath;
+      }
+    }
+  }
+  return EResourceLocation::kNotFound;
+}
+
+const void* LoadWinResource(const char* resid, const char* type, int& sizeInBytes, void* pHInstance)
+{
+  WDL_String typeUpper(type);
+
+  HRSRC hResource = FindResource(pHInstance, resid, _strupr(typeUpper.Get()));
+
+  if (!hResource)
+    return NULL;
+
+  DWORD size = SizeofResource(pHInstance, hResource);
+
+  if (size < 8)
+    return NULL;
+
+  HGLOBAL res = LoadResource(pHInstance, hResource);
+
+  const void* pResourceData = LockResource(res);
+
+  if (!pResourceData)
+  {
+    sizeInBytes = 0;
+    return NULL;
+  }
+  else
+  {
+    sizeInBytes = size;
+    return pResourceData;
+  }
+}
+
 #elif defined OS_WEB
 
 void AppSupportPath(WDL_String& path, bool isSystem)
@@ -159,6 +215,42 @@ void DesktopPath(WDL_String& path)
 void VST3PresetsPath(WDL_String& path, const char* mfrName, const char* pluginName, bool isSystem)
 {
   path.Set("Presets");
+}
+
+#include <emscripten/val.h>
+
+using namespace emscripten;
+
+EResourceLocation OSFindResource(const char* name, const char* type, WDL_String& result, const char*)
+{
+  if (CStringHasContents(name))
+  {
+    WDL_String plusSlash;
+    
+    bool foundResource = false;
+    
+    //TODO: OSFindResource is not sufficient here
+    
+    if(strcmp(type, "png") == 0) { //TODO: lowercase/uppercase png
+      plusSlash.SetFormatted(strlen("/resources/img/") + strlen(name) + 1, "/resources/img/%s", name);
+      foundResource = val::global("Module")["preloadedImages"].call<bool>("hasOwnProperty", std::string(plusSlash.Get()));
+    }
+    else if(strcmp(type, "ttf") == 0) { //TODO: lowercase/uppercase ttf
+      plusSlash.SetFormatted(strlen("/resources/fonts/") + strlen(name) + 1, "/resources/fonts/%s", name);
+      foundResource = true; // TODO: check ttf
+    }
+    else if(strcmp(type, "svg") == 0) { //TODO: lowercase/uppercase svg
+      plusSlash.SetFormatted(strlen("/resources/img/") + strlen(name) + 1, "/resources/img/%s", name);
+      foundResource = true; // TODO: check svg
+    }
+    
+    if(foundResource)
+    {
+      result.Set(plusSlash.Get());
+      return EResourceLocation::kAbsolutePath;
+    }
+  }
+  return EResourceLocation::kNotFound;
 }
 
 #endif
